added payment_collection check

+++ backend/src/jobs/cleanup-orphan-carts.ts
@@ -15,7 +15,7 @@ export default async function cleanupOrphanCarts(container: MedusaContainer) {
     // fetch carts with created at older than 24 hours and completed_at and deleted_at are null
     const { data: carts } = await query.graph({
       entity: Modules.CART,
-      fields: ['id', 'completed_at', 'deleted_at', 'created_at', 'order.id'],
+      fields: ['id', 'completed_at', 'deleted_at', 'created_at', 'order.id', 'payment_collection.id'],
       filters: {
         completed_at: null,
         deleted_at: null,
@@ -25,11 +25,15 @@ export default async function cleanupOrphanCarts(container: MedusaContainer) {
       },
     })
 
-    // fetch carts that are not associated with an order
+    // fetch carts that are not associated with an order and no payment collection available
     if (carts.length > 0) {
-      const cartIds = carts.filter(cart => !cart.order).map(cart => cart.id)
-      logger.info(`Deleting ${cartIds.length} orphan carts: ${cartIds.join(', ')}`)
-      await cartModuleService.deleteCarts(cartIds)
+      const cartIds = carts.filter(cart => !cart.order && !cart.payment_collection).map(cart => cart.id)
+      if (cartIds.length > 0) {
+        logger.info(`Deleting ${cartIds.length} orphan carts: ${cartIds.join(', ')}`)
+        await cartModuleService.deleteCarts(cartIds)
+      } else {
+        logger.info('No orphan carts found to delete.')
+      }
     } else {
       logger.info('No orphan carts found to delete.')
     }
